move integerattribute to manager
